Restructure i/o of data to support uproot and root backends
