+++ .gitignore
@@ -3,3 +3,7 @@
 *.pdf
 
 *.png
+
+.ipynb_checkpoints/
+
+*.DS_Store
